Successfully migrate emulator's IP and SP to scratch registers
Previously these used r12 and r13, which need to be preserved in the SysV ABI.
They have been moved to r8 and r9, which necessitated moving any scratch use
of r8 to use rsi instead.
The return code has also been moved to a scratch register, and a fourth has
been reserved for tracking the cycle count as instructions are run.