Reverse some previous work, move emulation state out of scratch registers
Finally remembered why they were originally placed in non-scratch registers...
Each re-compiled block is likely to call into external functions more
frequently than it itself is called. This means it is cheaper to push/pop some
non scratch registers at the beginning and end of the compiled block, than it
is to push/pop all scratch registers before entring a method like the one used
to call into memory.
This oversight was revealed when the SP register kept getting corrupted on a
specific test ROM.

Compilation itself will need to be reconfigured as well. Right now it's built
under the assumption that ROM is the only place in memory that can be
executed, and that it is a continuous unbanked slice of bytes. Some ROMs put
small routines in Work RAM or High RAM, so all code areas need to be
compilable. This is starting to be (messily) scaffolded out. Ideally the
cache will replace a single key (IP) with a multi-element tuple key
(Section, Bank, IP).

In other news, getting better debugging in place. "dump_dissassembly" feature
will print out the disassembly of each block that is compiled, giving some
indication of what happened before a crash. Future improvements will intercept
SIGSEGV signals and try to dump state and JIT'd code to a file on crash.